Buttons - Anna
All button scripts made and attached.

The canvas created for the death scene isnt working properly - you cant interact with the buttons, so i dont know if those buttons work.

All button code is in the ButtonController within each scene

Script that brings death screen up in death needs to be written and to look into the canvas - if it doesnt work ill make it as a separate scene