Updated to scenes and game loop
Updated the Game so that the player is racing against the clock to try and finish the level, added more lights and changed Anna's death canvas to be a scene instead and when the timer runs out, the scene changes, also fixed the buttons on the scenes from not working. Now everything works as intended.