Completed scenes and buttons
All buttons are functioning to switch to respected scenes.
Win screen was created and buttons were made and scrips for play and menu were attached.
Edited the collider script on the MICROWAVE so when the egg collides with it the win scene appears.
Edited Death screen so it has the egg texture.
Imported the egg model from maya - its got its own folder called egg models (who would have guessed) which also has the textures